use double instead of float

+++ Box2D/Box2D/Collision/Shapes/b2ChainShape.cpp
@@ -183,11 +183,11 @@ void b2ChainShape::ComputeAABB(b2AABB* aabb, const b2Transform& xf, int32_t chil
 	aabb->upperBound = b2Max(v1, v2);
 }
 
-void b2ChainShape::ComputeMass(b2MassData* massData, float density) const
+void b2ChainShape::ComputeMass(b2MassData* massData, double density) const
 {
 	B2_NOT_USED(density);
 
-	massData->mass = 0.0f;
+	massData->mass = 0.0;
 	massData->center.SetZero();
-	massData->I = 0.0f;
+	massData->I = 0.0;
 }
